+++ packages/editor/src/store/selectors.js
@@ -326,9 +326,9 @@ export function getEditedPostAttribute( state, attributeName ) {
  * Returns an attribute value of the current autosave revision for a post, or
  * null if there is no autosave for the post.
  *
- * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId )` selector from the
- *             '@wordpress/core-data' package and access properties on the returned autosave object
- *             using getPostRawValue.
+ * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId, userId )` selector
+ * 			   from the '@wordpress/core-data' package and access properties on the returned
+ * 			   autosave object using getPostRawValue.
  *
  * @param {Object} state         Global application state.
  * @param {string} attributeName Autosave attribute name.
@@ -337,7 +337,7 @@ export function getEditedPostAttribute( state, attributeName ) {
  */
 export const getAutosaveAttribute = createRegistrySelector( ( select ) => ( state, attributeName ) => {
 	deprecated( '`wp.data.select( \'core/editor\' ).getAutosaveAttribute( attributeName )`', {
-		alternative: '`wp.data.select( \'core\' ).getAutosave( postType, postId )`',
+		alternative: '`wp.data.select( \'core\' ).getAutosave( postType, postId, userId )`',
 		plugin: 'Gutenberg',
 	} );
 
@@ -347,7 +347,8 @@ export const getAutosaveAttribute = createRegistrySelector( ( select ) => ( stat
 
 	const postType = getCurrentPostType( state );
 	const postId = getCurrentPostId( state );
-	const autosave = select( 'core' ).getAutosave( postType, postId );
+	const currentUserId = get( select( 'core' ).getCurrentUser(), [ 'id' ] );
+	const autosave = select( 'core' ).getAutosave( postType, postId, currentUserId );
 
 	if ( autosave ) {
 		return getPostRawValue( autosave[ attributeName ] );
@@ -531,7 +532,8 @@ export const isEditedPostAutosaveable = createRegistrySelector( ( select ) => fu
 
 		const postType = getCurrentPostType( state );
 		const postId = getCurrentPostId( state );
-		autosave = select( 'core' ).getAutosave( postType, postId );
+		const currentUserId = get( select( 'core' ).getCurrentUser(), [ 'id' ] );
+		autosave = select( 'core' ).getAutosave( postType, postId, currentUserId );
 	}
 
 	// A post must contain a title, an excerpt, or non-empty content to be valid for autosaving.
@@ -563,7 +565,7 @@ export const isEditedPostAutosaveable = createRegistrySelector( ( select ) => fu
  * has yet to be autosaved, or has been saved or published since the last
  * autosave).
  *
- * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId )`
+ * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId, userId )`
  * 			   selector from the '@wordpress/core-data' package.
  *
  * @param {Object} state Editor state.
@@ -572,20 +574,21 @@ export const isEditedPostAutosaveable = createRegistrySelector( ( select ) => fu
  */
 export const getAutosave = createRegistrySelector( ( select ) => ( state ) => {
 	deprecated( '`wp.data.select( \'core/editor\' ).getAutosave()`', {
-		alternative: '`wp.data.select( \'core\' ).getAutosave( postType, postId )`',
+		alternative: '`wp.data.select( \'core\' ).getAutosave( postType, postId, userId )`',
 		plugin: 'Gutenberg',
 	} );
 
 	const postType = getCurrentPostType( state );
 	const postId = getCurrentPostId( state );
-	const autosave = select( 'core' ).getAutosave( postType, postId );
+	const currentUserId = get( select( 'core' ).getCurrentUser(), [ 'id' ] );
+	const autosave = select( 'core' ).getAutosave( postType, postId, currentUserId );
 	return mapValues( pick( autosave, AUTOSAVE_PROPERTIES ), getPostRawValue );
 } );
 
 /**
  * Returns the true if there is an existing autosave, otherwise false.
  *
- * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId )` selector
+ * @deprecated since 5.3. Callers should use the `getAutosave( postType, postId, userId )` selector
  *             from the '@wordpress/core-data' package and check for a truthy value.
  *
  * @param {Object} state Global application state.
@@ -594,13 +597,14 @@ export const getAutosave = createRegistrySelector( ( select ) => ( state ) => {
  */
 export const hasAutosave = createRegistrySelector( ( select ) => ( state ) => {
 	deprecated( '`wp.data.select( \'core/editor\' ).hasAutosave()`', {
-		alternative: '`!! wp.data.select( \'core\' ).getAutosave( postType, postId )`',
+		alternative: '`!! wp.data.select( \'core\' ).getAutosave( postType, postId, userId )`',
 		plugin: 'Gutenberg',
 	} );
 
 	const postType = getCurrentPostType( state );
 	const postId = getCurrentPostId( state );
-	return !! select( 'core' ).getAutosave( postType, postId );
+	const currentUserId = get( select( 'core' ).getCurrentUser(), [ 'id' ] );
+	return !! select( 'core' ).getAutosave( postType, postId, currentUserId );
 } );
 
 /**